# backend/src/astro_canvas/server/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Astro Canvas",
    "version": "0.1.0a0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/outputs/{node_id}/{port}": {
      "get": {
        "description": "A finished node's output. ``decimate``/``range`` apply to 1-d spectrum-like values.",
        "operationId": "get_output_api_outputs__node_id___port__get",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "port",
            "required": true,
            "schema": {
              "title": "Port",
              "type": "string"
            }
          },
          {
            "description": "Workflow the node belongs to.",
            "in": "query",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "description": "Workflow the node belongs to.",
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "fmt",
            "required": false,
            "schema": {
              "default": "json",
              "enum": [
                "json",
                "msgpack",
                "arrow",
                "npz"
              ],
              "title": "Fmt",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "decimate",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 4,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Decimate"
            }
          },
          {
            "description": "``lo,hi`` on the x axis.",
            "in": "query",
            "name": "range",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "``lo,hi`` on the x axis.",
              "title": "Range"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Output",
        "tags": [
          "outputs"
        ]
      }
    },
...
  }
}